Give five of `win_T`'s flags the `bool` they always were
These are the conversions a thawed layout unlocks and nothing else does: an
`int` that only ever holds 0 or 1 costs three bytes of padding under a C
layout and says nothing about its own range. All five are window state that
no option owns, which is why they are reachable at all — an option-owned
field stays a raw scalar whatever its struct's repr, because the option
table hands out its address as a `void *`.

- `w_set_curswant` (`int`, 79 sites, 63 of them a bare `= 1`)
- `w_hl_needs_update` (`int`)
- `w_arg_idx_invalid` (`int`, two sites already spelling `c_int::from(bool)`)
- `w_topline_was_set` (`char`)
- `w_stl_empty` (`char`, and the `empty_line` local feeding it, which was
  a `c_int::from` of a comparison and is now the comparison)

Each gets the doc comment it never had. The rest of `win_T`'s non-option
fields are named in the slice notes: `w_redr_type` is an ordered family of
`UPD_*` constants with 358 references tree-wide, and the nullable pointers
(`w_lines`, `w_alist`, `w_llist`, `w_localdir`, `w_ns_hl_attr`) each want a
sweep of their own.

+++ crates/nvim/src/arglist/mod.rs
@@ -779,7 +779,7 @@ pub unsafe fn check_arg_idx(win: *mut win_T) {
     if wargcount(win) <= 1 || editing {
         // Editing the current entry: `arg_had_last` if it is the last.
         // SAFETY: caller contract.
-        unsafe { (*win).w_arg_idx_invalid = c_int::from(false) };
+        unsafe { (*win).w_arg_idx_invalid = false };
         if idx == wargcount(win) - 1 && win_alist(win) == global_arglist() {
             arg_had_last.set(true);
         }
@@ -788,7 +788,7 @@ pub unsafe fn check_arg_idx(win: *mut win_T) {
     // Not editing the current entry, so `arg_had_last` only if this buffer
     // is the *last* global argument.
     // SAFETY: caller contract.
-    unsafe { (*win).w_arg_idx_invalid = c_int::from(true) };
+    unsafe { (*win).w_arg_idx_invalid = true };
     let gcount = alist_count(global_arglist());
     if idx == wargcount(win) - 1
         || arg_had_last.get()
